Quadratic Approximation for Profile Likelihood (Internal)

Approximates the information matrix via finite differences by perturbing beta in each dimension, re-profiling over y, and computing the difference in the log-likelihood.

Usage

quad.profile2(
  n,
  Hf,
  beta,
  y,
  delta,
  gamma,
  index,
  Z,
  eps = 1e-04,
  maxiter = 100
)

Arguments

n

Sample size, used to scale perturbations ptb = 1/sqrt(n).

Hf

A list of functions (H, H1, H2) defining the baseline hazard.

beta

Numeric vector of regression coefficients.

y

Numeric vector of baseline parameters (updated by profile1).

delta, gamma

Numeric 0/1 vectors indicating left and interval censoring status.

index

A 2-column integer matrix mapping each subject's interval to indices in y.

Z

A matrix or data frame of covariates.

eps

Convergence tolerance for the iterative algorithm.

maxiter

Maximum number of iterations to perform.

Value

A p x p matrix, an approximation to the negative Hessian (information).